mutual_fund_broker_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
requests
python-dotenv
apscheduler
